requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
twikit==2.3.1
